+++ src/main/kotlin/com/github/nvelychenko/drupalextend/type/EntityStorageTypeProvider.kt
@@ -1,17 +1,16 @@
 package com.github.nvelychenko.drupalextend.type
 
+import com.github.nvelychenko.drupalextend.extensions.getValue
 import com.github.nvelychenko.drupalextend.extensions.isSuperInterfaceOf
 import com.github.nvelychenko.drupalextend.index.ContentEntityIndex
+import com.github.nvelychenko.drupalextend.patterns.Patterns
 import com.github.nvelychenko.drupalextend.type.EntityStorageTypeProvider.Util.SPLITER_KEY
-import com.github.nvelychenko.drupalextend.util.getValue
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
-import com.intellij.patterns.PlatformPatterns
 import com.intellij.psi.PsiElement
+import com.intellij.psi.util.childrenOfType
 import com.intellij.util.indexing.FileBasedIndex
 import com.jetbrains.php.PhpIndex
-import com.jetbrains.php.lang.parser.PhpElementTypes
-import com.jetbrains.php.lang.psi.PhpPsiUtil
 import com.jetbrains.php.lang.psi.elements.*
 import com.jetbrains.php.lang.psi.resolve.types.PhpType
 import com.jetbrains.php.lang.psi.resolve.types.PhpTypeProvider4
@@ -22,6 +21,10 @@ import java.util.zip.GZIPInputStream
 import java.util.zip.GZIPOutputStream
 
 
+/**
+ * $entity_storage = \Drupal::entityTypeManager()->getStorage('node');
+ *       ↑
+ */
 class EntityStorageTypeProvider : PhpTypeProvider4 {
 
     private val entityTypeManagerInterface = "\\Drupal\\Core\\Entity\\EntityTypeManagerInterface"
@@ -30,40 +33,27 @@ class EntityStorageTypeProvider : PhpTypeProvider4 {
         return Util.KEY
     }
 
-    override fun getType(psiElement: PsiElement): PhpType? {
-        if (DumbService.getInstance(psiElement.project).isDumb) {
-            return null
-        }
-
-        // container calls are only on "get" methods
-        if (psiElement !is MethodReference) {
-            return null
-        }
-
-        if (psiElement.name != "getStorage") {
-            return null
-        }
-
-        if (!PlatformPatterns
-                .psiElement(PhpElementTypes.METHOD_REFERENCE)
-                .withChild(
-                    PlatformPatterns
-                        .psiElement(PhpElementTypes.PARAMETER_LIST)
-                        .withFirstChild(
-                            PlatformPatterns.psiElement(StringLiteralExpression::class.java),
-                        )
-                ).accepts(psiElement)
+    override fun getType(methodReference: PsiElement): PhpType? {
+        if (
+            DumbService.getInstance(methodReference.project).isDumb
+            || !Patterns.METHOD_WITH_FIRST_STRING_PARAMETER.accepts(methodReference)
         ) {
             return null
         }
 
-        val parameterList = PhpPsiUtil.getChildOfClass(psiElement, ParameterList::class.java)
-        if (parameterList !is ParameterList) return null
+        methodReference as MethodReference
+
+        if (methodReference.name != "getStorage") {
+            return null
+        }
+
+        val parameterList = methodReference.childrenOfType<ParameterList>().first()
 
         val firstParameter = parameterList.getParameter(0) as StringLiteralExpression
+
         if (firstParameter.contents.isEmpty()) return null
 
-        val signature = compressString(psiElement.signature).replace('|', SPLITER_KEY)
+        val signature = compressString(methodReference.signature).replace('|', SPLITER_KEY)
         return PhpType().add("#$key${signature}${Util.SPLIT_KEY}${firstParameter.contents}")
     }
 
@@ -80,7 +70,48 @@ class EntityStorageTypeProvider : PhpTypeProvider4 {
     }
 
     override fun complete(expression: String?, project: Project?): PhpType? {
-        return null
+        if (project == null || expression == null || !expression.contains(Util.SPLIT_KEY)) return null
+
+        val signature = expression.replace("#${key}", "")
+        val parts = signature.split(Util.SPLIT_KEY)
+
+        if (parts.size != 2) {
+            return null
+        }
+
+        val (originalSignature, entityTypeId) = parts
+
+        val entityType =
+            FileBasedIndex.getInstance().getValue(ContentEntityIndex.KEY, entityTypeId, project) ?: return null
+
+        val phpIndex = PhpIndex.getInstance(project)
+        val namedCollection = mutableListOf<PhpNamedElement>()
+        for (partialSignature in decompressString(originalSignature.replace(SPLITER_KEY, '|')).split('|')) {
+            namedCollection.addAll(phpIndex.getBySignature(partialSignature))
+        }
+
+        val methods = namedCollection.filterIsInstance<Method>()
+
+        if (methods.isEmpty()) return null
+
+        val entityTypeManagerInterface =
+            phpIndex.getInterfacesByFQN(entityTypeManagerInterface).takeIf { it.isNotEmpty() }?.first()
+                ?: return null
+
+        val entityStorageClass = phpIndex.getClassesByFQN(entityType.storageHandler).takeIf { it.isNotEmpty() }?.first()
+            ?: return null
+
+        val type = PhpType()
+        methods.forEach { method ->
+            val clazz = method.containingClass ?: return@forEach
+            if (clazz.fqn == entityTypeManagerInterface.fqn || clazz.isSuperInterfaceOf(entityTypeManagerInterface)) {
+                entityStorageClass.implementsList.referenceElements.map {
+                    it.fqn
+                }.forEach(type::add)
+            }
+        }
+
+        return type
     }
 
     override fun getBySignature(
@@ -88,66 +119,14 @@ class EntityStorageTypeProvider : PhpTypeProvider4 {
         visited: Set<String?>?,
         depth: Int,
         project: Project?
-    ): Collection<PhpNamedElement> {
-        // @todo Figure out why id doesn't work, here should be this.getBySignature method that should work
-        //   but it brakes ContentEntityFieldTypeProvider type provider.
-        return emptyList()
-    }
-
-    private fun getBySignature(
-        project: Project?,
-        expression: String,
-        visited: Set<String?>?,
-        depth: Int
-    ): Collection<PhpNamedElement> {
-        if (project == null || !expression.contains(Util.SPLIT_KEY)) return emptyList()
-
-        val parts = expression.split(Util.SPLIT_KEY)
-
-        if (parts.size != 2) {
-            return emptyList()
-        }
-
-        val (originalSignature, entityTypeId) = parts
-
-        val entityType =
-            FileBasedIndex.getInstance().getValue(ContentEntityIndex.KEY, entityTypeId, project) ?: return emptyList()
-
-        val phpIndex = PhpIndex.getInstance(project)
-        val namedCollection = mutableListOf<PhpNamedElement>()
-        for (partialSignature in decompressString(originalSignature.replace(SPLITER_KEY, '|')).split('|')) {
-            namedCollection.addAll(phpIndex.getBySignature(partialSignature, visited, depth))
-        }
-
-        val methods = namedCollection.filterIsInstance<Method>()
-
-        if (methods.isEmpty()) return emptyList()
-
-        val baseStorageClass = phpIndex.getClassesByFQN("\\Drupal\\Core\\Entity\\EntityStorageBase")
-
-        methods.forEach {
-            if (it.fqn == entityTypeManagerInterface) {
-                return phpIndex.getClassesByFQN(entityType.storageHandler) + baseStorageClass
-            }
-        }
-
-        val entityTypeManagerInterface =
-            phpIndex.getInterfacesByFQN(entityTypeManagerInterface).takeIf { it.isNotEmpty() }?.first()
-                ?: return emptyList()
-
-        methods.forEach {
-            if (it.containingClass?.isSuperInterfaceOf(entityTypeManagerInterface) == true) {
-                return phpIndex.getClassesByFQN(entityType.storageHandler) + baseStorageClass
-            }
-        }
-
-        return emptyList()
+    ): Collection<PhpNamedElement>? {
+        return null
     }
 
     object Util {
-        val SPLIT_KEY = '\u3333'
-        val KEY = '\u3334'
-        val SPLITER_KEY = '\u3336'
+        const val SPLIT_KEY = '\u3333'
+        const val KEY = '\u3334'
+        const val SPLITER_KEY = '\u3336'
     }