+++ src/main/kotlin/com/github/nvelychenko/drupalextend/type/FieldItemListTypeProvider.kt
@@ -17,9 +17,13 @@ import com.jetbrains.php.lang.psi.resolve.types.PhpTypeProvider4
 import org.apache.commons.lang3.StringUtils
 
 
-class FieldPropertyTypeProvider : PhpTypeProvider4 {
+/**
+ * $item_list = $entity->get('field_foo');
+ *     ↑
+ */
+class FieldItemListTypeProvider : PhpTypeProvider4 {
 
-    val endKey = '\u3339'
+    private val endKey = '\u3339'
 
     override fun getKey(): Char {
         return '\u3338'