+++ CHANGELOG.md
@@ -4,6 +4,54 @@
 
 ## [Unreleased]
 
+## [0.3.2] - 2024-01-31
+
+### Added
+
+- Autocomplete for render elements and their properties.
+- Line marker for Render/Form element
+- Completion priority for direct properties.
+- Implemented the ability to try to find entity from storage only knowing its interface/class.
+- Storage autocomplete in case Symfony is enabled, and storage is returned from service.
+
+## [0.3.1] - 2024-01-31
+
+### Added
+
+- Line marker for Render/Form element
+- Completion priority for direct properties.
+
+## [0.3.0] - 2024-01-30
+
+Added:
+
+- Implemented ability to try to find entity from storage only knowing its interface/class.
+- annotator bug
+- storage autocomplete in case Symfony is enabled, and storage is returned from service.
+
+## [0.2.0] - 2024-01-30
+
+### Added
+
+- Autocomplete for render elements and their properties.
+
+## [0.1.0] - 2024-01-28
+
+### Added
+
+- Integration with symfony service type provider for \Drupal::service
+- Config entity support (autocomplete, references)
+- Field properties autocomplete e.g. $node->get('field_address')->addr| (it will autocomplete address_line1, address_line2)
+- Node::load/create/loadMultiple type provider
+- Almost all base fields are autocompleted now. (fields specified in a few static methods on the same class WIP)
+- Storage autocomplete for getAccessControlHandler, getStorage, getViewBuilder, getListBuilder, getFormObject, getRouteProviders, hasHandler methods.
+- Type provider for storage again work! But now it provider only interfaces
+
+### Fixed
+
+- Fixed bug related to broken entities type provider.
+- Fields autocomplete in inappropriate places.
+
 ## [0.0.5]
 
 ### Added
@@ -19,5 +67,10 @@
 - Content entity provider was not working in certain circumstances
 - Deprecated usage of baseDir
 
-[Unreleased]: https://github.com/nvelychenko/drupal-extend/compare/0.0.5...HEAD
+[Unreleased]: https://github.com/nvelychenko/drupal-extend/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/nvelychenko/drupal-extend/compare/v0.3.1...v0.3.2
+[0.3.1]: https://github.com/nvelychenko/drupal-extend/compare/v0.3.0...v0.3.1
+[0.3.0]: https://github.com/nvelychenko/drupal-extend/compare/v0.2.0...v0.3.0
+[0.2.0]: https://github.com/nvelychenko/drupal-extend/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/nvelychenko/drupal-extend/compare/0.0.5...v0.1.0
 [0.0.5]: https://github.com/nvelychenko/drupal-extend/commits/0.0.5